requirements.txt
blas==1.0
blosc==1.21.0
brotli==1.0.9
bzip2==1.0.8
ca-certificates==2022.3.29
certifi==2020.6.20
cfitsio==3.470
charls==2.2.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
cytoolz==0.11.0
dask-core==2022.2.1
fonttools==4.31.2
freetype==2.10.4
fsspec==2022.2.0
giflib==5.2.1
icc_rt==2019.0.0
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
intel-openmp==2021.4.0
jpeg==9d
kiwisolver==1.4.2
lcms2==2.12
lerc==3.0
libaec==1.0.4
libdeflate==1.8
libffi==3.4.2
libpng==1.6.37
libtiff==4.2.0
libwebp==1.2.2
libzopfli==1.0.3
locket==0.2.1
lz4-c==1.9.3
matplotlib==3.5.1
mkl-service==2.4.0
mkl==2021.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
networkx==2.7.1
numpy-base==1.21.5
numpy==1.21.5
opencv-python==4.5.5.64
openjpeg==2.4.0
openssl==1.1.1n
packaging==21.3
partd==1.2.0
pillow==9.0.1
pip==22.0.4
pycantonese==3.4.0
pylangacq==0.16.2
pyparsing==3.0.4
python-dateutil==2.8.2
python==3.10.4
pywavelets==1.3.0
pyyaml==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
snappy==1.1.8
sqlite==3.38.2
tabulate==0.8.9
tifffile==2021.7.2
tk==8.6.11
toolz==0.11.2
tzdata==2022a
urllib3==1.26.9
vc==14.2
vs2015_runtime==14.27.29016
wcwidth==0.2.5
wheel==0.37.1
wincertstore==0.2
wordseg==0.0.2
xz==5.2.5
yaml==0.2.5
zfp==0.5.5
zlib==1.2.11
zstd==1.4.9
